pokemon updated/deleted does cahnge team.pokemon[]
If a pokemon gets updated, it also updates the pokemon in teams if its in a team.
If pokemon gets deleted, it also gets deleted from all existing teams if the pokemon is part of team(s).

twaeked clicks on button (was mostly on links and not the whole button)